[Refactor] File 구조 재정리 + useFile hook 생성
