+++ src/App.jsx
@@ -9,7 +9,7 @@ import Chatbot from './routes/Chatbot';
 import useReissue from './hooks/useReissue';
 import ChatProf from './routes/ChatProf';
 import Writing from './routes/Writing';
-import File from './routes/File';
+import File from './routes/File/File';
 import useShowAlarm from './hooks/useShowAlarm';
 
 function App() {
