+++ src/routes/File/File.css
@@ -82,6 +82,34 @@
 
 .title-col {
   flex: 0 0 50%;
+  display: flex;
+  align-items: center;
+}
+
+
+.title-text {
+  display: inline-block;
+  max-width: calc(100% - 50px); 
+  white-space: nowrap;
+  overflow: hidden; 
+  text-overflow: ellipsis; 
+}
+
+.new-badge {
+  display: inline-block;
+  right: 10px;
+  border: 1px solid rgba(10, 124, 247, 1);
+  color: rgba(10, 124, 247, 1);
+  padding: 2px 3px;
+  font-size: 12px;
+  border-radius: 5px;
+  height: 18px;
+  width: 30px;
+  line-height: 18px;
+  height: 18px;
+  text-align: center;
+  margin-left: 5px;
+  margin-right: 5px;
 }
 
 .date-col {
